// lib/types.ts
export type JobCreator = {
  uid: string;
  name: string;
  employeeId?: string;
};

export type Job = {
  id: string;
  title: string;
  from: string;
  to:string;
  pickupCoordinates: { lat: number; lng: number };
  vehicleType: 'Truck' | 'Van' | 'Car';
  status: 'Awaiting' | 'In Progress' | 'Completed';
  requirements: {
    location: string;
    vehicleType: string;
    licenseValidity: boolean;
    insuranceValidity: boolean;
  };
  createdBy: JobCreator;
  createdAt?: string; // ISO 8601 string format
  bookingTime?: string; // ISO 8601 string for scheduled time
  customerName?: string;
  paymentMethod?: 'Invoice' | 'Card' | 'Cash';
  paymentStatus?: 'Paid' | 'Unpaid';
};
// ...
export type VehicleJob = {
    id: string | null;
    status: string;
    service: string | null;
    account: string | null;
    pickup: string | null;
    pickupDate: string | null;
    bookingTime?: string; // ISO 8601 string for scheduled time
    flight: string | null;
    pax: string | null;
    distance: number;
    dwellTime: string;
    eta: string | null;
    plannedRoute: LatLngExpression[] | null;
    actualRoute: LatLngExpression[] | null;
    trip?: {
        id: number;
        status: string;
        jobs: string[];
    }
};

export type Vehicle = {
  id: string;
  callsign: string;
  licensePlate: string;
  vehicleType: string;
  make: string;
  model: string;
  capacity: string;
  status: 'On Duty' | 'Idle' | 'Maintenance' | 'Offline' | 'On Break';
  position: LatLngExpression;
  driver: {
    id: string;
    name: string;
    phone: string;
    earnings: number;
  };
  job: VehicleJob;
  telemetry: {
    fuel: number; // percentage
    battery: number; // percentage
  };
  breakInfo?: {
    startTime: string;
    endTime: string;
  };
};
// ...
export type Break = {
    startTime: string;
    endTime: string;
    duration: string;
    driverCallsign: string;
};
// ...
export type Rate = {
    id: string;
    vehicleType: string;
    category: 'passenger' | 'shipper' | 'equipment';
    percentPer: number;
    rate: number;
    vatPercent: number;
    vatAmount: number;
    damages: number;
    handling: number;
    waiting: number;
    halting: number;
};